Restriction in elimination of HumanResource

+++ src/main/java/com/apms/humanResource/HumanResourceRestController.java
@@ -15,7 +15,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.apms.rest.RESTRequest;
 import com.apms.rest.RESTResponse;
+
 import com.apms.user.UserService;
+import com.apms.user.User;
 
 @RestController
 @RequestMapping("/humanResource")
@@ -121,13 +123,13 @@ public class HumanResourceRestController {
     @DeleteMapping("/{id}")
     public RESTResponse<HumanResource> delete(@PathVariable Integer id) {
         try {
-            //List<User> users = userService.getUsersByHumanResource(id);
-            //if (users.isEmpty()) {
+            User auxUser = userService.getUserByHumanResource(id);
+            if (auxUser == null) {
                 humanResourceService.delete(id);
-            //} else {
-            //    return new RESTResponse<HumanResource>(RESTResponse.FAIL,
-            //            "Este Recurso Humano esta asociado a un Usuario por lo que no puede ser eliminado.", null);
-            //}
+            } else {
+               return new RESTResponse<HumanResource>(RESTResponse.FAIL,
+                       "Este Recurso Humano esta asociado a un Usuario por lo que no puede ser eliminado.", null);
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
